test(tui): isolate skill label localization assertion from web summary tests
